+++ megalinter/flavor_factory.py
@@ -35,7 +35,7 @@ def list_megalinter_flavors():
             "label": "Optimized for CI items (Dockerfile, Jenkinsfile, JSON/YAML schemas, XML)"
         },
         "dart": {"label": "Optimized for DART based projects"},
-        "documentation": {"label": "Mega-Linter for documentation projects"},
+        "documentation": {"label": "Optimized for documentation projects"},
         "dotnet": {"label": "Optimized for C, C++, C# or VB based projects"},
         "go": {"label": "Optimized for GO based projects"},
         "java": {"label": "Optimized for JAVA based projects"},